migration and entities

+++ src/author.rs
@@ -1,7 +1,9 @@
+use crate::schema::author;
+
 #[table_name = "author"]
 #[derive(AsChangeset, Serialize, Deserialize, Queryable, Insertable)]
 pub struct Author {
-    pub id: Option<i64>,
+    pub id: Option<i32>,
     pub name: String,
     pub username: String,
     pub resume: String,
